refactor: move DetailRoomInfo.tsx to Modal folder move file location
